Use issuer's SignatureAlgorithm for CRL signing
Presumably the issuer's SignatureAlgorithm was chosen for a reason. With
PSS support, this could mean the underlying key is using a PKCS#11
mechanism. Alternatively, a longer hash function was chosen, aligning
with some other requirements (ECDSA size, &c).

Rather than leaving it up to the Go library to choose signature
algorithm, proactively mirror it from the issuer's certificate.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>